+++ pluggable-architecture/git-stats/tests/plugins/cassettes/test_bitbucket_plugin.yaml
